feat(service-template): rename SERVICE_NAME->service_name because docker needs lowercase service name
