--- SERVICE_NAME/auth/__init__.py
@@ -1,1 +1,0 @@
-from authutils.user import current_user
