+++ deployment/nginx/uwsgi.conf
@@ -6,7 +6,7 @@ server {
         uwsgi_param REMOTE_USER $http_x_userid if_not_empty;
 
         include uwsgi_params;
-        uwsgi_pass unix:/var/www/SERVICE_NAME/uwsgi.sock;
+        uwsgi_pass unix:/var/www/service_name/uwsgi.sock;
         uwsgi_read_timeout 20s;
         uwsgi_send_timeout 20s;
     }
@@ -16,7 +16,7 @@ server {
         uwsgi_param REMOTE_USER $http_x_userid if_not_empty;
 
         include uwsgi_params;
-        uwsgi_pass unix:/var/www/SERVICE_NAME/uwsgi.sock;
+        uwsgi_pass unix:/var/www/service_name/uwsgi.sock;
         uwsgi_read_timeout 20s;
         uwsgi_send_timeout 20s;
         access_log off;
