+++ deployment/uwsgi/uwsgi.ini
@@ -1,19 +1,19 @@
 [uwsgi]
 protocol = uwsgi
-socket = /var/www/SERVICE_NAME/uwsgi.sock
+socket = /var/www/service_name/uwsgi.sock
 buffer-size = 32768
 chmod-socket = 666
 master = true
 processes = 2
 harakiri-verbose = 20
 disable-logging = true
-wsgi-file=/var/www/SERVICE_NAME/wsgi.py
+wsgi-file=/var/www/service_name/wsgi.py
 plugins = python
 vacuum = true
 uid = www-data
 gid = www-data
-pythonpath = /var/www/SERVICE_NAME/
-pythonpath = /SERVICE_NAME/
+pythonpath = /var/www/service_name/
+pythonpath = /service_name/
 
 # Initialize application in worker processes, not master. This prevents the
 # workers from all trying to open the same database connections at startup.
