+++ service_name/admin_endpoints/__init__.py
@@ -1,7 +1,7 @@
 from flask import Blueprint
 
-from SERVICE_NAME import auth
-from SERVICE_NAME.errors import JWTError
+from service_name import auth
+from service_name.errors import JWTError
 
 
 blueprint = Blueprint('admin', __name__)
@@ -40,8 +40,11 @@ def check_if_user_is_admin():
         403:
             description: User is not admin
     """
-    # check if user is admin (raises error if user is not connected)
-    if auth.current_user.is_admin:
-        return 'Success! User is admin', 200
-    else:
-        return 'Woops! User is not admin', 403
+    try:
+        # check if user is admin (raises error if user is not connected)
+        if auth.current_user.is_admin:
+            return 'Success! User is admin', 200
+        else:
+            return 'Woops! User is not admin', 403
+    except JWTError as e:
+        return e.message, e.code
