+++ service_name/auth/__init__.py
@@ -1,0 +1,2 @@
+# from authutils import ROLES
+from authutils.user import current_user
